+++ app/winapi.go
@@ -58,6 +58,8 @@ const (
 	wmSyskeyup       = 0x105
 	vkTab            = 0x09
 	vkControl        = 0x11
+	vkMenu           = 0x12
+	vkEnd            = 0x23
 	vkLwin           = 0x5B
 	vkRwin           = 0x5C
 	vkSnapshot       = 0x2C